requirements.txt
discord.py
async-timeout==3.0.1
python-dotenv